#7455 - BE - Remit Reconciliation Process
The remittance query was running slow in QA, so I made the query more efficient. The count query was removed, and other query will return all rows not just the current page. The count can be determined from the query result (size of list). The requested page is inserted in the returned list. The time to run this process has been reduced to half or less.

Signed-off-by: Erik Krason <erik.krason@mcmcg.com>