#15354 - BE - Create UserLoginHistory DAO Implementation #15355 - BE - Create UserLoginHistory Query
New DAO code for UserLoginHistory. When I pushed changes for Application code I forgot to format the text. I ran this code through the text formatter, so there are a lot of changes. I tested the new code in Postman and it correctly inserts a row in the database for login.

Signed-off-by: Erik Krason <erik.krason@mcmcg.com>